Add test for sleep
